Fix Safari auto-play issues
