+++ src/triage/patient-set.tsx
@@ -5,35 +5,16 @@ import {
 
 interface PatientSetProps {
     patients: Patient[];
+    stopAllAudio(): void;
+    playAudio(url: string): void;
 }
 
-export const PatientSet: React.FC<PatientSetProps> = ({patients}: PatientSetProps) => {
+export const PatientSet: React.FC<PatientSetProps> = (
+    {patients, stopAllAudio, playAudio}: PatientSetProps) => {
     const [currentPatient, setCurrentPatient] = useState<number | null>(null);
     const [countdownClock, setCountdownClock] = useState<number | null>(null);
     const [lockPanel, setLockPanel] = useState<boolean>(false);
     const interval = useRef<number | null>(null);
-    const audioRef = useRef<HTMLAudioElement | null>(null);
-
-    const stopAllAudio = (): void => {
-        if (audioRef.current) {
-            audioRef.current.pause();
-            audioRef.current = null;
-        }
-    };
-
-    const playAudio = (audioURL: string): void => {
-        // Stop all audio
-        stopAllAudio();
-        // Start the first audio
-        const a = new Audio();
-        // eslint-disable-next-line scanjs-rules/call_addEventListener
-        a.addEventListener('canplaythrough', () => {
-            void a.play();
-        });
-        // eslint-disable-next-line scanjs-rules/assign_to_src
-        a.src = audioURL;
-        audioRef.current = a;
-    };
 
     const countdown = (): void => {
         setCountdownClock((prev) => {
@@ -52,7 +33,14 @@ export const PatientSet: React.FC<PatientSetProps> = ({patients}: PatientSetProp
         setLockPanel(false);
         setCurrentPatient(idx);
         setCountdownClock(Number(patients[idx].countdown));
-        playAudio(patients[idx].promptAudio);
+        // When starting the sim, the initial play event needs to occur
+        // when there's a click event. For idx == 0, the audio is started
+        // by the parent Triage component
+        // This particular corner case is to satify Safari's autoplay policies
+        // https://webkit.org/blog/6784/new-video-policies-for-ios/
+        if (idx > 0) {
+            playAudio(patients[idx].promptAudio);
+        }
 
         // Now start timer, and countdown clock
         // eslint-disable-next-line scanjs-rules/call_setInterval
